Allows operator input right after pair of numbers.
A pair of numbers accepted only equal input and therefore did not
support operator input immediately after a pair of numbers.
Added features which allows operator input after a pair of numbers
as well. Achieved by checking if b input is existing, if yes, the
calculator operates for an answer, then update inputArr and display.