Avoid adding digits to first number after operation
To fix a bug which user being stuck with adding digits to the first
number, given the previous number pair is operated. The calculator
should start with a new first number with digit input after an operation,
instead of continuing adding digits to the first number.

To avoid this bug, add a condition to check if a and answer are entered,
while b and op are not. If condition met, clearAll() to let user start
with empty variable inputs.